업무 로직 변경 오류 수정

+++ model/userModel/blacklistModel.js
@@ -1,7 +1,7 @@
 let database = require('../../database/database');
 let querybuilder = require('../../util/queryBuilder');
-let tablename = 'black_user';
-let userModel = '../../model/';
+let tablename = 'blacklist';
+let userModel = require('./userModel');
 let logger = require('../../util/logger');
 
 const user_id = "user_id",
@@ -22,16 +22,16 @@ function addUserToBlackList(param_user_id, dataObj, callback) {
 // 유저 테이블로 가서 블랙리스트 정보에 필요한 유저정보를 선별해서 가져와야 함
 function getMyBlackList(param_user_id, callback) {
     logger.debug('[3]blacklistDao-getMyBlackList');
-    let query = `SELECT * FROM ${tablename} WHERE ${user_id}=${param_user_id}`;
+    let query = `SELECT * FROM ${tablename} WHERE ${user_id}='${param_user_id}'`;
     database.executeByRaw(query, (err, data)=>{
-        let userIdList = querybuilder.build(data, id);
+        let userIdList = querybuilder.build(data, black_id, id);
         userModel.getUserInfo(userIdList, callback);
     });
 }
 
-function removeUserFromBlackList(param_black_id, dataObj, callback) {
+function removeUserFromBlackList(param_user_id, param_black_id, dataObj, callback) {
     logger.debug('[3]blacklistDao-removeUserFromBlackList');
-    let query = `DELETE FROM ${tablename} WHERE ${black_id}=${param_black_id}`
+    let query = `DELETE FROM ${tablename} WHERE ${user_id}='${param_user_id}' AND ${black_id}='${param_black_id}'`
     database.executeByRaw(query, callback);
 }
